refactor(gas): move simulation to lotus-shed, clean up test file
- Move the 3600-epoch gas market simulation from gasutils_test.go to
  cmd/lotus-shed/gas-market-sim.go as a proper CLI tool with --epochs,
  --seed, and --epoch-remaining-zero flags replacing env vars.

- Clean up gasutils_test.go: remove dead timestamp parameter from
  makeTipSet (always 0), widen receiver to testing.TB so the benchmark
  can use it, replace the 30-iteration convergence logging loop with a
  direct assertion, and remove the inline tipset construction from the
  benchmark.
